fix : merge main.rs and /binmain.rs in backend
